Move some functions into utils.ts
