+++ src/esmodule.ts
@@ -4,6 +4,8 @@ import {
     TarDefaults,
     TarEntryType,
 } from "./constants.js";
+import { generateChecksum } from "./utils.js";
+import { normalizePath, splitFilename, isZeroed } from "./utils.js";
 
 export class TarArchive implements Iterable<TarEntry> {
     /** A map of file paths to entry indicies. */
@@ -467,49 +469,6 @@ export class TarArchive implements Iterable<TarEntry> {
     }
 }
 
-/** Normalizes a path. */
-function normalizePath(path: string): string {
-    const entries = path.split("/");
-
-    for (let i = 0; i < entries.length; i++) {
-        switch (entries[i]) {
-            case "":
-            case ".":
-                entries.splice(i, 1);
-                i--;
-                break;
-
-            case "..":
-                if (i >= 1) {
-                    entries.splice(i - 1, 2);
-                    i -= 2;
-                }
-                break;
-
-            default:
-                break;
-        }
-    }
-
-    return entries.join("/");
-}
-
-/** Splits filename into a prefix and name. */
-function splitFilename(path: string) {
-    if (path.length > 255) {
-        throw new Error(`Path is too long: ${path.length} > 255`);
-    }
-
-    let _name = path;
-    let _prefix = "";
-    if (path.length > 100) {
-        let sepIndex = path.length - 100;
-        _name = path.substring(sepIndex);
-        _prefix = path.substring(0, sepIndex);
-    }
-    return { _name, _prefix };
-}
-
 function tarHeader(
     name: string,
     mode: number,
@@ -630,23 +589,6 @@ function writeOctal(
     }
 }
 
-/** Generate a checksum of an array of bytes with a given precision. */
-function generateChecksum(bytes: Uint8Array, precision: number): number {
-    const mask = (2 << precision) - 1;
-    const chksumStart = 0x094;
-    const chksumEnd = 0x09b;
-
-    let value = 0;
-    for (let i = 0; i < bytes.length; i++) {
-        let byte = bytes[i];
-        if (chksumStart <= i && i <= chksumEnd) {
-            byte = 0x20;
-        }
-        value = (value + byte) & mask;
-    }
-    return value;
-}
-
 /** Decodes a null-terminated string from an array of bytes. */
 function decodeString(
     bytes: Uint8Array,
@@ -686,11 +628,6 @@ function decodeOctal(
         value |= byte - 0x30;
     }
     return value;
-}
-
-/** Returns whether or not all bytes are zero. */
-function isZeroed(bytes: Uint8Array): boolean {
-    return bytes.every((v) => v === 0);
 }
 
 export interface TarEntry {
